Add PartialWarningsCompile
Works like Compile but adds three new arguments: partialwarningStart,
partialwarningEnd, partialwarningProperty which makes warnings not to
be collected until a line that matches the partialwarningStart regexp
is found. Collection stops when partialwarningEnd matches. In addition
to storing the collected warnings into a separate logfile the property
given by partialwarningProperty is updated with the count.